backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ML & Computer Vision
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
pillow==10.1.0
opencv-python==4.9.0.80
numpy==1.26.2
grad-cam==1.4.8

# Storage
supabase==2.3.0
storage3==0.7.0

# Utilities
pydantic==2.5.3
aiofiles==23.2.1
requests==2.31.0
